refactor: split holdings-grid.tsx into folder with sub-components
Closes #366
